Deprecate TR_ARMOpCode
This class can now be replaced by `OMR::ARM::InstOpCode`. We simply copy
over all the APIs from the former to the latter. There is no change in
functionality following this commit.

Further we remove OpBinary.cpp and OpProperties.cpp and replace it with
OMRInstOpCode.cpp.
